pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "smallpond"
version = "0.15.0"
description = "A lightweight data processing framework built on DuckDB and shared file system."
authors = [
    { name = "DeepSeek-AI", email = "research@deepseek.com" },
    { name = "Runji Wang" },
    { name = "Yiliang Xiong" },
    { name = "Yiyuan Liu" },
    { name = "Yuheng Zou" },
    { name = "Yichao Zhang" },
    { name = "Wenjun Gao" },
    { name = "Wentao Zhang" },
    { name = "Xiaotao Nie" },
    { name = "Minghua Zhang" },
    { name = "Zhewen Hao" },
]
urls = { Homepage = "https://github.com/deepseek-ai/smallpond" }
keywords = ["distributed query processing", "SQL", "parquet"]
requires-python = ">=3.8"
dependencies = [
    "duckdb >= 1.2.0",
    "pyarrow ~= 16.1.0",
    "polars ~= 0.20.9",
    "pandas >= 1.3.4",
    "plotly >= 5.22.0",
    "lxml >= 4.9.3",
    "cloudpickle >= 2.0.0",
    "zstandard >= 0.22.0",
    "loguru >= 0.7.2",
    "psutil >= 5.9.8",
    "GPUtil >= 1.4.0",
    "py-libnuma >= 1.2",
    "fsspec >= 2023.12.2",
    "ray[default] >= 2.10.0",
    "graphviz >= 0.19.1",
]

[project.optional-dependencies]
dev = [
    "coverage~=7.4.4",
    "hypothesis~=6.100.0",
    "pytest==8.2.1",
    "pytest-cov==5.0.0",
    "pytest-forked==1.6.0",
    "pytest-xdist==3.6.1",
    "pytest-timeout==2.3.1",
    "pytest-benchmark==4.0.0",
    "setproctitle==1.3.3",
    "soupsieve~=2.5",
    "setuptools-scm==8.1.0",
    "packaging==24.2",
    "jaraco.functools==4.1.0",
]
docs = [
    "sphinx==7.1.2",
    "pydata-sphinx-theme==0.14.4",
]
warc = [
    "warcio >= 1.7.4",
    "beautifulsoup4 >= 4.12.2",
]

[tool.setuptools]
packages = [
    "smallpond",
    "smallpond.io",
    "smallpond.execution",
    "smallpond.logical",
    "smallpond.contrib",
    "smallpond.platform",
]

[tool.setuptools_scm]
fallback_version = "0.0.0"
